ce_expr.yy now handles projections and selections using 'value'
This code removes the value struct from the ce_expr parser
and improves processing selection expressions by removing
unneeded casting operations that were, in some cases, breaking
valid CEs (because negative values were 'floored' to zero when
comparing them to a value in an unsigned variable. The old code
is still present in this commit - wrapped in #if 0 ... #endif.